feat: increment or default version (#16)
We should increment OR default the version on server-side.

This is applicable in cases where are are multiple callers trying to
upsert something.
Previously, PynamoDB will always set the version attribute to 1 if the
caller does not explicitly set it.
Imagine the scenario:
Worker 1 wants to upsert an entry (id=123, name=Jake,
timestamp=10:00:00)
Worker 2 wants to upsert an entry (id=123, name=Jake,
timestamp=11:00:00)
The caller will always need to explicitly state an action that says
version=version + 1 or 1 if missing (otherwise it will "upsert" without
actually incrementing the version).

---------

Co-authored-by: Jinying Li <ji20@snapchat.com>